+++ rollup.config.mjs
@@ -1,8 +1,5 @@
 import commonjs from '@rollup/plugin-commonjs';
 import nodeResolve from '@rollup/plugin-node-resolve';
-import babel from '@rollup/plugin-babel';
-import '@rollup/plugin-terser';
-import json from '@rollup/plugin-json';
 import terser from '@rollup/plugin-terser';
 
 const dev = process.env.ROLLUP_WATCH;
@@ -10,12 +7,6 @@ const dev = process.env.ROLLUP_WATCH;
 const plugins = [
   nodeResolve({}),
   commonjs(),
-  json(),
-  babel({
-    exclude: 'node_modules/**',
-    babelHelpers: 'bundled',
-    babelrc: false,
-  }),
   !dev &&
   terser({
     format: {